move lambda script to it's own folder with it's own package.json so we can easily create a zip file of the index file and node_modules folder
