+++ .gitignore
@@ -65,3 +65,4 @@ public/participedia-data-organizations.csv
 
 .idea/
 
+aws-lambda-functions/optimize-images.zip
