+++ package.json
@@ -143,7 +143,7 @@
     }
   },
   "lint-staged": {
-    "*.js": [
+    "*/*.js": [
       "prettier --write",
       "git add"
     ]
